Complete porting of classes and tests
