+++ src/Generator/ASTFactory/functions.php
@@ -2,5 +2,12 @@
 declare(strict_types=1);
 
 require_once __DIR__ . '/createClass.php';
-require_once __DIR__ . '/getVisibility.php';
+require_once __DIR__ . '/createMethod.php';
+require_once __DIR__ . '/createMethodGet.php';
+require_once __DIR__ . '/createMethodCall.php';
+require_once __DIR__ . '/createParameter.php';
+require_once __DIR__ . '/createParameterDefaultValue.php';
 require_once __DIR__ . '/createProperty.php';
+require_once __DIR__ . '/createPropertyUnset.php';
+require_once __DIR__ . '/createReturnType.php';
+require_once __DIR__ . '/getVisibility.php';
